Cleanup BraceAnalyzerPlugin Project
