+++ ExperimentingWithPluginSystem/BraceAnalyzerPlugin/BraceAnalyzerPlugin.sln
@@ -1,0 +1,16 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BraceAnalyzerPlugin", "BraceAnalyzerPlugin\BraceAnalyzerPlugin.csproj", "{CE44F2BF-6653-4CE2-B3B3-9ABFF93DDFE5}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{CE44F2BF-6653-4CE2-B3B3-9ABFF93DDFE5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CE44F2BF-6653-4CE2-B3B3-9ABFF93DDFE5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CE44F2BF-6653-4CE2-B3B3-9ABFF93DDFE5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CE44F2BF-6653-4CE2-B3B3-9ABFF93DDFE5}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
