+++ ExperimentingWithPluginSystem/BraceAnalyzerPlugin/BraceAnalyzerPlugin/Resources.resx
@@ -19,7 +19,7 @@
         <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=2.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
     </resheader>
     <data name="BraceAnalyzerDescription" xml:space="preserve">
-        <value>Enclose statement with curly braces</value>
+        <value>Enclose statement with curly braces.</value>
     </data>
     <data name="BraceAnalyzerMessageFormat" xml:space="preserve">
         <value>'{' brace expected</value>
