--- ExperimentingWithPluginSystem/PluginTest2/PluginTest2.sln
@@ -1,16 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PluginTest2", "PluginTest2.csproj", "{58644918-DC92-4FB1-A2AF-8084D422ED68}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{58644918-DC92-4FB1-A2AF-8084D422ED68}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{58644918-DC92-4FB1-A2AF-8084D422ED68}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{58644918-DC92-4FB1-A2AF-8084D422ED68}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{58644918-DC92-4FB1-A2AF-8084D422ED68}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
